Move the expandPermutations helper to its own file
