+++ lib/util/fonts/extractTextFromContentPropertyValue.js
@@ -1,6 +1,7 @@
 var unescapeCssString = require('./unescapeCssString');
 var charactersByListStyleType = require('./charactersByListStyleType');
 var getCounterCharacters = require('./getCounterCharacters');
+var expandPermutations = require('../expandPermutations');
 
 function extractQuotes(quotes, token) {
     if (!quotes || quotes === 'none') {
@@ -88,27 +89,6 @@ function expandContent(tokens, node, quotes, hypotheticalCounterStyleByName) {
     return text;
 }
 
-function expandHypotheticalCounterStylePermutations(hypotheticalCounterStyleValuesByName, referencedCounterStyleNames) {
-    var permutations = [];
-    var firstPropertyName = referencedCounterStyleNames[0];
-    var firstPropertyValues = hypotheticalCounterStyleValuesByName[referencedCounterStyleNames[0]];
-
-    for (var i = 0 ; i < Math.max(1, firstPropertyValues.length) ; i += 1) {
-        if (referencedCounterStyleNames.length > 1) {
-            expandHypotheticalCounterStylePermutations(hypotheticalCounterStyleValuesByName, referencedCounterStyleNames.slice(1)).forEach(function (permutation) {
-                permutation[firstPropertyName] = firstPropertyValues[i];
-                permutations.push(permutation);
-            });
-        } else {
-            var permutation = {};
-            permutation[firstPropertyName] = firstPropertyValues[i];
-            permutations.push(permutation);
-        }
-    }
-
-    return permutations;
-}
-
 function extractTextFromContentPropertyValue(value, node, hypotheticalQuotesValues, hypotheticalCounterStyleValuesByName) {
     var tokens = tokenizeContent(value);
     var isSeenByCounterStyle = {};
@@ -138,7 +118,7 @@ function extractTextFromContentPropertyValue(value, node, hypotheticalQuotesValu
         var referencedCounterStyleNames = Object.keys(isSeenByCounterStyle);
         if (referencedCounterStyleNames.length > 0) {
             var result = [];
-            expandHypotheticalCounterStylePermutations(hypotheticalCounterStyleValuesByName, referencedCounterStyleNames).forEach(function (hypotheticalCounterStyleByName) {
+            expandPermutations(hypotheticalCounterStyleValuesByName, referencedCounterStyleNames).forEach(function (hypotheticalCounterStyleByName) {
                 var content = expandContent(tokens, node, quotes, hypotheticalCounterStyleByName);
                 if (content) {
                     result.push({
