+++ lib/util/fonts/getTextByFontProperties.js
@@ -12,6 +12,7 @@ var extractTextFromContentPropertyValue = require('./extractTextFromContentPrope
 var charactersByListStyleType = require('./charactersByListStyleType');
 var unescapeCssString = require('./unescapeCssString');
 var getCounterCharacters = require('./getCounterCharacters');
+var expandPermutations = require('../expandPermutations');
 
 var PROPS = [
     'font-family',
@@ -40,28 +41,6 @@ var INITIAL_VALUES = {
 var NAME_CONVERSIONS = {
     'font-weight': cssFontWeightNames
 };
-
-// Expand into all permutations in case of multiple hypothetical values:
-function expandPermutations(styledText, propertyNames) {
-    propertyNames = propertyNames || Object.keys(styledText);
-    var permutations = [];
-    var firstPropertyName = propertyNames[0];
-    var firstPropertyValues = styledText[propertyNames[0]];
-
-    for (var i = 0 ; i < Math.max(1, firstPropertyValues.length) ; i += 1) {
-        if (propertyNames.length > 1) {
-            expandPermutations(styledText, propertyNames.slice(1)).forEach(function (permutation) {
-                permutation[firstPropertyName] = firstPropertyValues[i];
-                permutations.push(permutation);
-            });
-        } else {
-            var permutation = {};
-            permutation[firstPropertyName] = firstPropertyValues[i];
-            permutations.push(permutation);
-        }
-    }
-    return permutations;
-}
 
 function createPredicatePermutations(predicatesToVary, i) {
     if (typeof i !== 'number') {
